Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.16",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "mem0ai>=0.1.116",
    "pypdf>=6.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.11.0",
]
